extract player to a wrapper class
